fix: don't truncate 64-bit EVM chain IDs in the ETH signing UI
The ETH transaction parser (Rust) correctly reads the chain ID as a u64,
but the C UI layer printed it with "%lu" (GetEthNetWork). On the device
target (thumbv7em, 32-bit `long`) this only consumes the low 32 bits, so
any chain ID >= 2^32 that is not in the built-in network table is shown
incorrectly. The simulator (64-bit host) hides the bug.

Example: glamsterdam-devnet-8 has chain ID 7091047534 (0x1A6A8CC6E,
33 bits); the device displayed 2796080238 (the low 32 bits), which looks
like a plausible-but-different chain to the user signing.

"%llu"/PRIu64 is not a fix here because the firmware links newlib-nano
(--specs=nano.specs), whose formatted IO has no long long support.
Add a small Uint64ToDecStr helper and use it instead.

Also fix two related 32-bit narrowings on the same path:
- EIP-712 domain chainId lookup used atoi() (int) -> strtoull()
- GetDBContract() took the chain ID as uint32_t -> uint64_t
  (firmware + simulator stub)

+++ src/user_sqlite3.h
@@ -15,6 +15,6 @@
 void Sqlite3Test(int argc, char *argv[]);
 void UserSqlite3Init(void);
 bool GetEnsName(const char *addr, char *name);
-bool GetDBContract(const char* address, const char *selector, const uint32_t chainId, char *functionABIJson, char *contractName);
+bool GetDBContract(const char* address, const char *selector, const uint64_t chainId, char *functionABIJson, char *contractName);
 
 #endif
